requirements.txt
# ── Etapa 1: Ingestão e Pré-processamento ──
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyarrow>=14.0
matplotlib>=3.7
seaborn>=0.12

# ── Etapa 2: RAG ──
sentence-transformers>=2.2
chromadb>=0.4
pyyaml>=6.0

# ── Interface (será usada nas etapas seguintes) ──
streamlit>=1.30

# ── Utilidades ──
python-dotenv>=1.0

# ── Testes ──
pytest>=7.4

# ── Futuro: LLM ──
# ollama>=0.1
